Enable user to preview their own pfp, show link-copied's style after copying previewed image's link

+++ src/components/Chat/Message/MessageDelete.jsx
@@ -29,7 +29,9 @@ export function MessageDelete({ closeModal, conversationId, deletingData }) {
 
         const imagesDocs = await getDocs(imagesQueryRef);
 
-        await deleteDoc(imagesDocs.docs[0].ref);
+        if (imagesDocs.docs.length) {
+          await deleteDoc(imagesDocs.docs[0].ref);
+        }
         
       }
